sdk: rename sdk scripts lib functions to avoid confusion with ebuilds
It makes no sense for the utility library functions of the SDK scripts
to have nearly the same names as the utility functions used within the
ebuilds. This can only lead to confusion for new developers and
prevents us to deviate from the function behavior of the Gentoo eponym
functions.

Change-Id: I96421116e55bd49d5457ceceee62eb0d76a7b534
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>