sdk: setup-portage: create global package.{,un}mask & license from profile
Setup package.mask, package.unmask & package.license in /etc/portage/
using special *.global-override folders from overlay profiles.

This is a workaround for a Portage limitation that requires those files
to be in /etc/portage and will not read them in profiles overlays or do
not allow some operations from overlays.

Change-Id: I2571c58b58a9b30a689558071b42b53e718cedb2
Signed-off-by: Timothée Ravier <timothee.ravier@ssi.gouv.fr>