sdk: build.sh: emerge clipos-kernel-sources by default
We deliberatly build `sys-kernel/clipos-kernel-sources` during the
"build" step (but not on the "image" step as we do not want to have
the kernel sources in the final images) in order not to break the
potential kernel configuration settings checks that some ebuilds may
implement with the use of the "check_extra_config" function provided
by the "linux-info" eclass. (e.g. systemd ebuild which checks that all
the absolutely required kernel configuration settings are set to make
run properly systemd).

If we do not populate the rootfs (during "build" step time) with the
configured sources of the CLIP OS kernel, then the function
"check_extra_config" fail and report misleading error messages on
emerge output (e.g. telling that a CONFIG_FOO setting is not set in
the kernel where it should be).

Change-Id: Ic9aa23416f0b6bd466a1adbf7a08b51f7273643d
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>