core: configure: Alternative implementation for root password setting when instrumented
This version is safer because it is more agnostic on the preliminary
content of the /etc/{passwd,shadow} files it works on. This version is
also able to raise a fatal error when the process for setting an empty
root password and /bin/bash as the default shell fails (rather than
failing silently).

Change-Id: I879ea32e70b66250d7f16bab673d6922966ca5d1
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>