sdk: setup-portage: Do not set LINGUAS as unadvised by Gentoo docs
Prefer to keep LINGUAS unset (as it is by default) because LINGUAS has
the same behaviour in Portage as gettext defines it.
This means that when LINGUAS is unset, all the available localization
files are compiled but not necessarily installed (this is controlled
by the L10N setting). When LINGUAS is set but empty, no localization
file is compiled (and therefore, nor installed). When set and defined
to a value, only matching localization files are compiled.
We would like to prevent unnecessary packages rebuilds when the target
localization change by only leaving the L10N setting install the
appropriate localization files.

See:
  - https://wiki.gentoo.org/wiki/Localization/Guide#LINGUAS
  - https://www.gentoo.org/support/news-items/2016-06-23-l10n-use_expand.html

Change-Id: I349df53b122b0461dd716c239a06f28539f3d60c
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>