sdk: configure scripts: Revamp locale generation to avoid requiring a shell
As locale-gen is a Bash script, this commit rewrites the locales
generation to directly use the binary "localdef" (provided by the
glibc with the locales) in order to avoid the need to emerge a shell
interpreter and basic command-line tools (i.e. sed and Awk) in the
Core rootfs.

This locale generation method also checks that the locales have been
properly generated and gives a feedback to the CLIP OS developer in
the configure scripts logs.

Change-Id: I926fe585f40e64db6b32952349b73f38bb42d8dc
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>