efiboot: configure: Only use serial console in debug instru
This will avoid people with no TPM emulator installed complaining about
not being prompted for the passphrase to decrypt the disk, as it is
actually done in the text console and not in the graphical one. Plus
it makes more sense anyway now that we can SSH into the VM for
development.

Change-Id: Ib42d9903eba5aa18e4391b13853eebbdb9ddc5c1
Signed-off-by: Thibaut Sautereau <thibaut.sautereau@ssi.gouv.fr>