efiboot: Exclude PCR 4 when sealing state secret
The firmware will extend the value in PCR 4 with the measurement of EFI
binaries loaded by the bootloader (systemd-boot). This value will be
different for each CLIP OS version and is thus excluded from the list of
PCRs used to seal the state secret.

See section "7.2.4 Measuring OS Boot Events" in:
https://trustedcomputinggroup.org/wp-content/uploads/PC-ClientSpecific_Platform_Profile_for_TPM_2p0_Systems_v51.pdf

Change-Id: If31c8edc5468eedb2816accf8c6b030e2cf821f4
Signed-off-by: Timothée Ravier <timothee.ravier@ssi.gouv.fr>