core/configure: Prepare /etc/hosts for local resolver
There is a need for custom host entries at runtime for configuration
update by administrators. As there is currently no mechanism to expand
the static /etc/hosts file, we keep it in a minimal form.

Temporary entries used for development and testing will be replaced
later by the integration of a local DNS resolver for each network zone.

Change-Id: Iaddf86b36125be4d3de1c8cb7a2453cfa7aac99b
Signed-off-by: Timothée Ravier <timothee.ravier@ssi.gouv.fr>